Move PaymentSource to @type/
