+++ src/core/providers/payment/StripeProvider.ts
@@ -7,7 +7,7 @@ import { CompletedPaymentFlow } from "@core/providers/payment-flow";
 
 import stripe from "@core/lib/stripe";
 import { log as mainLogger } from "@core/logging";
-import { getActualAmount, PaymentForm, PaymentSource } from "@core/utils";
+import { getActualAmount, PaymentForm } from "@core/utils";
 import { calcRenewalDate, getChargeableAmount } from "@core/utils/payment";
 import {
   createSubscription,
@@ -22,7 +22,9 @@ import { StripePaymentData } from "@models/PaymentData";
 import NoPaymentMethod from "@api/errors/NoPaymentMethod";
 
 import config from "@config";
+
 import { ContributionInfo } from "@type/contribution-info";
+import { PaymentSource } from "@type/payment-source";
 
 const log = mainLogger.child({ app: "stripe-payment-provider" });
