+++ src/core/utils/index.ts
@@ -1,4 +1,4 @@
-import { ContributionPeriod, PaymentMethod } from "@beabee/beabee-common";
+import { ContributionPeriod } from "@beabee/beabee-common";
 import { NextFunction, Request, RequestHandler, Response } from "express";
 import { QueryFailedError } from "typeorm";
 
@@ -8,47 +8,6 @@ export interface PaymentForm {
   payFee: boolean;
   prorate: boolean;
 }
-
-export interface GoCardlessDirectDebitPaymentSource {
-  method: PaymentMethod.GoCardlessDirectDebit;
-  bankName: string;
-  accountHolderName: string;
-  accountNumberEnding: string;
-}
-
-export interface StripeCardPaymentSource {
-  method: PaymentMethod.StripeCard;
-  last4: string;
-  expiryMonth: number;
-  expiryYear: number;
-}
-
-export interface StripeBACSPaymentSource {
-  method: PaymentMethod.StripeBACS;
-  sortCode: string;
-  last4: string;
-}
-
-export interface StripeSEPAPaymentSource {
-  method: PaymentMethod.StripeSEPA;
-  country: string;
-  bankCode: string;
-  branchCode: string;
-  last4: string;
-}
-
-export interface ManualPaymentSource {
-  method: null;
-  source?: string;
-  reference?: string;
-}
-
-export type PaymentSource =
-  | GoCardlessDirectDebitPaymentSource
-  | StripeCardPaymentSource
-  | StripeBACSPaymentSource
-  | StripeSEPAPaymentSource
-  | ManualPaymentSource;
 
 export function getActualAmount(
   amount: number,
