+++ src/core/utils/payment/stripe.ts
@@ -8,10 +8,12 @@ import Stripe from "stripe";
 
 import stripe from "@core/lib/stripe";
 import { log as mainLogger } from "@core/logging";
-import { PaymentForm, PaymentSource } from "@core/utils";
+import { PaymentForm } from "@core/utils";
 import { getChargeableAmount } from "@core/utils/payment";
 
 import config from "@config";
+
+import { PaymentSource } from "@type/payment-source";
 
 const log = mainLogger.child({ app: "stripe-utils" });
